requirements.txt

numpy==1.24.3
pandas==2.0.3
torch==2.0.1
transformers==4.31.0
tiktoken==0.5.1
openai==1.1.1
python-dotenv==1.0.0
flask==2.3.3
cryptography==41.0.3
diffusers==0.21.4
accelerate==0.23.0
pillow==10.0.0
tqdm==4.66.1
scikit-learn==1.3.0
